Update lots of code for 2.0.0 test with XT merges
- Added new visual aids for: melee attacks, supply station noises, and func_buttons
- Added some checks in animation system and ways to track down errors going forward
- Fixed up legwound effect to play correctly
- Made talking balloon smaller than other floating sprites
- Prevent spam score commands make it a queue
- Remove suicide delay message for 0 delay, prevent negative delays
- Stopping following a player in spectator now cleans up better by moving back to spectator and restoring position
- Added g_balancedDeathAmmo def 1 ( 25 sh, 25 na, 10 rk, 50 ce )
- Moved g_spawnFullStats to 1 even though full stats are always happening now can remove later
- Added g_newPulseGren def 0, testing new Pulse grenade damage on player behavior but it is not finished
- Needle stab and wrench now count towards humiliation
- Local client now properly initializes as admin and rcon in menus
- 200 cells are now required to upgrade supply station per level. Be careful not to fill up your supply station if you do not wish to fill it when attacking with wrench.
- Pulse grenades now can cause napalm rockets to explode
- Added support for delayfirstspawn key on goalinfos for delayed spawns like classic powerups work.
- Demo menu now supports folders
- Increase maximum number of menus, demos, configs, huds
- Increase overall UI menu memory pool availability
- Added additional menu defines for versioning for main menu

+++ code/cgame/cg_q3f_spawn.c
@@ -520,7 +520,7 @@ static char *CG_Q3F_AddSpawnVarToken( const char *string ) {
 	int		l;
 	char	*dest;
 
-	l = strlen( string );
+	l = (int)strlen( string );
 	if ( numSpawnVarChars + l + 1 > MAX_SPAWN_VARS_CHARS ) {
 		CG_Error( "CG_Q3F_AddSpawnVarToken: MAX_SPAWN_VARS_CHARS" );
 	}
